Feat:新增完整的软件包下载、安装的逻辑，这部分在整个构建过程中通用

+++ src/cross_compile_script/libstdc++.sh
@@ -1,4 +1,6 @@
-mkdir -v build && cd build 
+mkdir -v build 
+cd build 
+
 ../libstdc++-v3/configure           \
     --host=$LFS_TGT                 \
     --build=$(../config.guess)      \
@@ -7,5 +9,14 @@ mkdir -v build && cd build
     --disable-nls                   \
     --disable-libstdcxx-pch         \
     --with-gxx-include-dir=/tools/$LFS_TGT/include/c++/12.2.0
+
+
+if [ "$?" -eq 1 ];
+then
+    exit $?
+fi
+
 make && make DESTDIR=$LFS install
+
+
 rm -v $LFS/usr/lib/lib{stdc++,stdc++fs,supc++}.la
